First commit, added source and comments to README
Added the source to the project, most of it was taken from adafruit but
modified to work with the max7219 and LedControl library. Currently
using a fast fourier transform library, but when focusing on the
software side i will most likely switch to a fast hartley transform
algorithm. Filled in README with more useful information as well.